Move os_log streaming from output channel to task terminal
Replace the singleton LogStreamManager (which wrote to a separate
"SweetPad: App Logs" output channel) with per-launch LoggingManager
instances that write directly to the build task terminal. Default
device log backend changed from "off" to "osActivityDtMode".
