+++ package.json
@@ -970,12 +970,12 @@
         "sweetpad.build.logStreamEnabled": {
           "type": "boolean",
           "default": true,
-          "description": "Enable log streaming during debug sessions. Captures os_log/Logger/NSLog output in the 'SweetPad: App Logs' output channel. print() output is also captured via stdout."
+          "description": "Enable os_log/Logger streaming in the task terminal during app launch. When enabled, os_log output appears alongside print() and build output in the same terminal."
         },
         "sweetpad.build.logStreamPredicate": {
           "type": "string",
           "default": null,
-          "description": "Custom predicate for log stream filtering. Use ${bundleId} for the bundle identifier and ${processName} for the process name. Default: captures os_log/Logger (by subsystem) and NSLog (via Foundation sender).",
+          "description": "Custom predicate for log stream filtering. Use ${bundleId} for the bundle identifier and ${processName} for the process name. Default: captures os_log/Logger by subsystem.",
           "examples": [
             "subsystem BEGINSWITH \"${bundleId}\" OR (process == \"${processName}\" AND sender == \"Foundation\")",
             "subsystem BEGINSWITH \"${bundleId}\"",
@@ -989,13 +989,13 @@
             "osActivityDtMode",
             "pymobiledevice3"
           ],
-          "default": "off",
+          "default": "osActivityDtMode",
           "enumDescriptions": [
-            "Do not stream os_log from physical devices (default).",
+            "Do not stream os_log from physical devices.",
             "Inject OS_ACTIVITY_DT_MODE=enable so os_log mirrors to stderr via devicectl --console. Zero external dependencies, requires Xcode 16+. Undocumented Apple behavior, may break on future iOS versions. Mirrors main-executable logs only; dynamically-loaded framework logs may be missing.",
             "Spawn `pymobiledevice3 syslog live` to stream structured os_log (subsystem, category, level). Requires `pip install pymobiledevice3` and, on iOS 17+, a running tunnel (`sudo pymobiledevice3 remote tunneld`)."
           ],
-          "description": "Backend for streaming os_log/Logger output from physical iOS devices during debug sessions."
+          "description": "Backend for streaming os_log/Logger output from physical iOS devices during app launch."
         },
         "sweetpad.build.pymobiledevice3Path": {
           "type": "string",
