+++ src/logging/filters.spec.ts
@@ -1,10 +1,9 @@
 import {
-  PassthroughLogPipe,
-  Pymobiledevice3JsonLogPipe,
-  Pymobiledevice3LogPipe,
-  buildAppImageFilter,
+  PassthroughLogFilter,
+  Pymobiledevice3JsonLogFilter,
+  Pymobiledevice3LogFilter,
   parseSyslogLine,
-} from "./log-pipe";
+} from "./filters";
 
 describe("parseSyslogLine", () => {
   it("parses a line with subsystem and category (the --label suffix)", () => {
@@ -77,43 +76,20 @@ describe("parseSyslogLine", () => {
   });
 });
 
-describe("buildAppImageFilter", () => {
-  const make = (imageName: string) => ({
-    timestamp: "t",
-    processName: "Laboratory",
-    imageName,
-    pid: 1,
-    level: "NOTICE",
-    message: "m",
+describe("PassthroughLogFilter", () => {
+  it("passes non-empty lines through", () => {
+    const filter = new PassthroughLogFilter();
+    expect(filter.processLine("hello world")).toBe("hello world");
   });
 
-  it("keeps both main image and debug dylib by default", () => {
-    const keep = buildAppImageFilter("Laboratory");
-    expect(keep(make("Laboratory"))).toBe(true);
-    expect(keep(make("Laboratory.debug.dylib"))).toBe(true);
-  });
-
-  it("debugDylibOnly=true keeps only the debug dylib", () => {
-    const keep = buildAppImageFilter("Laboratory", true);
-    expect(keep(make("Laboratory.debug.dylib"))).toBe(true);
-    expect(keep(make("Laboratory"))).toBe(false);
-  });
-
-  it("drops framework-emitted noise (CoreFoundation, Foundation, etc.)", () => {
-    const keep = buildAppImageFilter("Laboratory");
-    expect(keep(make("CoreFoundation"))).toBe(false);
-    expect(keep(make("Foundation"))).toBe(false);
-    expect(keep(make("libdispatch.dylib"))).toBe(false);
-  });
-
-  it("does not match by prefix — 'LaboratoryExt' is a different image", () => {
-    const keep = buildAppImageFilter("Laboratory");
-    expect(keep(make("LaboratoryExt"))).toBe(false);
+  it("drops empty lines", () => {
+    const filter = new PassthroughLogFilter();
+    expect(filter.processLine("")).toBeNull();
   });
 });
 
-describe("Pymobiledevice3LogPipe", () => {
-  function createPipe(
+describe("Pymobiledevice3LogFilter", () => {
+  function createFilter(
     options: {
       executableName: string;
       debugDylibOnly?: boolean;
@@ -122,89 +98,79 @@ describe("Pymobiledevice3LogPipe", () => {
       minLevel?: string;
     } = { executableName: "Laboratory" },
   ) {
-    const lines: string[] = [];
-    const pipe = new Pymobiledevice3LogPipe((line) => lines.push(line), options);
-    return { pipe, lines };
+    return new Pymobiledevice3LogFilter(options);
   }
 
   it("drops framework noise", () => {
-    const { pipe, lines } = createPipe();
+    const filter = createFilter();
     const noise =
       "2026-04-16 12:52:32.707333 Laboratory{CoreFoundation}[67116] <NOTICE>: noise [com.apple.defaults][User Defaults]";
-    pipe.push(`${noise}\n`);
-    expect(lines).toEqual([]);
+    expect(filter.processLine(noise)).toBeNull();
   });
 
   it("keeps app logs from debug dylib", () => {
-    const { pipe, lines } = createPipe();
+    const filter = createFilter();
     const appLog =
       "2026-04-16 12:53:59.832625 Laboratory{Laboratory.debug.dylib}[67135] <NOTICE>: hello [Lab Test][App]";
-    pipe.push(`${appLog}\n`);
-    expect(lines).toEqual([appLog]);
+    expect(filter.processLine(appLog)).toBe(appLog);
   });
 
   it("keeps the main image by default (debugDylibOnly=false)", () => {
-    const { pipe, lines } = createPipe();
+    const filter = createFilter();
     const line = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: hi [sub][cat]";
-    pipe.push(`${line}\n`);
-    expect(lines).toEqual([line]);
+    expect(filter.processLine(line)).toBe(line);
   });
 
   it("debugDylibOnly=true drops the main image", () => {
-    const { pipe, lines } = createPipe({ executableName: "Laboratory", debugDylibOnly: true });
+    const filter = createFilter({ executableName: "Laboratory", debugDylibOnly: true });
     const mainImage = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <DEBUG>: noise [com.apple.Previews][X]";
     const debugDylib = "2026-04-16 12:53:59.832625 Laboratory{Laboratory.debug.dylib}[1] <NOTICE>: app log [sub][cat]";
-    pipe.push(`${mainImage}\n`);
-    pipe.push(`${debugDylib}\n`);
-    expect(lines).toEqual([debugDylib]);
+    expect(filter.processLine(mainImage)).toBeNull();
+    expect(filter.processLine(debugDylib)).toBe(debugDylib);
   });
 
   it("subsystem deny-list drops matching subsystems", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemDenyList: ["com.apple.*"],
     });
     const apple = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <DEBUG>: noise [com.apple.CFBundle][resources]";
     const app = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: hi [com.example.app][ui]";
-    pipe.push(`${apple}\n`);
-    pipe.push(`${app}\n`);
-    expect(lines).toEqual([app]);
+    expect(filter.processLine(apple)).toBeNull();
+    expect(filter.processLine(app)).toBe(app);
   });
 
   it("subsystem deny-list does not affect entries without a label", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemDenyList: ["com.apple.*"],
     });
     const noLabel = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: no label here";
-    pipe.push(`${noLabel}\n`);
-    expect(lines).toEqual([noLabel]);
+    expect(filter.processLine(noLabel)).toBe(noLabel);
   });
 
   it("subsystem allow-list keeps only matching subsystems", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemAllowList: ["com.example.*"],
     });
     const match = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: a [com.example.app][ui]";
     const other = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: b [com.apple.CFBundle][resources]";
-    pipe.push(`${match}\n`);
-    pipe.push(`${other}\n`);
-    expect(lines).toEqual([match]);
+    expect(filter.processLine(match)).toBe(match);
+    expect(filter.processLine(other)).toBeNull();
   });
 
   it("subsystem allow-list drops entries without a label", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemAllowList: ["com.example.*"],
     });
     const noLabel = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: no label";
-    pipe.push(`${noLabel}\n`);
-    expect(lines).toEqual([]);
+    expect(filter.processLine(noLabel)).toBeNull();
   });
 
   it("deny-list and allow-list work together", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemDenyList: ["com.apple.*"],
       subsystemAllowList: ["com.example.myapp"],
@@ -212,99 +178,57 @@ describe("Pymobiledevice3LogPipe", () => {
     const apple = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: a [com.apple.X][Y]";
     const myApp = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: b [com.example.myapp][ui]";
     const otherApp = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: c [com.example.other][ui]";
-    pipe.push(`${apple}\n`);
-    pipe.push(`${myApp}\n`);
-    pipe.push(`${otherApp}\n`);
-    expect(lines).toEqual([myApp]);
+    expect(filter.processLine(apple)).toBeNull();
+    expect(filter.processLine(myApp)).toBe(myApp);
+    expect(filter.processLine(otherApp)).toBeNull();
   });
 
   it("honors an optional minimum level", () => {
-    const { pipe, lines } = createPipe({ executableName: "Laboratory", minLevel: "ERROR" });
+    const filter = createFilter({ executableName: "Laboratory", minLevel: "ERROR" });
     const info = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <INFO>: quiet [s][c]";
     const error = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <ERROR>: loud [s][c]";
-    pipe.push(`${info}\n`);
-    pipe.push(`${error}\n`);
-    expect(lines).toEqual([error]);
+    expect(filter.processLine(info)).toBeNull();
+    expect(filter.processLine(error)).toBe(error);
   });
 
   it("passes through tunnel notices before the first parsed entry", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push("Connected to tunnel\n");
-    expect(lines).toEqual(["Connected to tunnel"]);
+    const filter = createFilter();
+    expect(filter.processLine("Connected to tunnel")).toBe("Connected to tunnel");
   });
 
   it("drops empty lines", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push("\n");
-    expect(lines).toEqual([]);
+    const filter = createFilter();
+    expect(filter.processLine("")).toBeNull();
   });
 
   it("drops continuation lines after a dropped entry", () => {
-    const { pipe, lines } = createPipe();
+    const filter = createFilter();
     const dropped =
       "2026-04-16 12:53:59.832625 Laboratory{CoreFoundation}[1] <DEBUG>: first [com.apple.CFBundle][resources]";
-    pipe.push(`${dropped}\n`);
-    pipe.push("    Localizations : [en]\n");
-    pipe.push("    Dev language  : en\n");
-    expect(lines).toEqual([]);
+    expect(filter.processLine(dropped)).toBeNull();
+    expect(filter.processLine("    Localizations : [en]")).toBeNull();
+    expect(filter.processLine("    Dev language  : en")).toBeNull();
   });
 
   it("keeps continuation lines after a kept entry", () => {
-    const { pipe, lines } = createPipe();
+    const filter = createFilter();
     const kept = "2026-04-16 12:53:59.832625 Laboratory{Laboratory.debug.dylib}[1] <NOTICE>: multi-line [sub][cat]";
-    pipe.push(`${kept}\n`);
-    pipe.push("    second line\n");
-    pipe.push("    third line\n");
-    expect(lines).toEqual([kept, "    second line", "    third line"]);
+    expect(filter.processLine(kept)).toBe(kept);
+    expect(filter.processLine("    second line")).toBe("    second line");
+    expect(filter.processLine("    third line")).toBe("    third line");
   });
 
   it("passes through unparseable lines before the first parsed entry (tunnel notices)", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push("Connected to tunnel\n");
-    pipe.push("Enabling developer disk image\n");
+    const filter = createFilter();
+    expect(filter.processLine("Connected to tunnel")).toBe("Connected to tunnel");
+    expect(filter.processLine("Enabling developer disk image")).toBe("Enabling developer disk image");
     const dropped = "2026-04-16 12:53:59.832625 Laboratory{CoreFoundation}[1] <DEBUG>: noise [com.apple.X][Y]";
-    pipe.push(`${dropped}\n`);
-    pipe.push("    continuation of noise\n");
-    expect(lines).toEqual(["Connected to tunnel", "Enabling developer disk image"]);
-  });
-
-  it("buffers partial fragments across chunks", () => {
-    const { pipe, lines } = createPipe();
-    const line = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: hi [s][c]";
-    pipe.push(line.slice(0, 10));
-    expect(lines).toEqual([]);
-    pipe.push(`${line.slice(10)}\n`);
-    expect(lines).toEqual([line]);
-  });
-
-  it("flush() emits a trailing fragment without newline", () => {
-    const { pipe, lines } = createPipe();
-    const line = "2026-04-16 12:53:59.832625 Laboratory{Laboratory}[1] <NOTICE>: hi [s][c]";
-    pipe.push(line);
-    expect(lines).toEqual([]);
-    pipe.flush();
-    expect(lines).toEqual([line]);
+    expect(filter.processLine(dropped)).toBeNull();
+    expect(filter.processLine("    continuation of noise")).toBeNull();
   });
 });
 
-describe("PassthroughLogPipe", () => {
-  it("forwards chunks as-is", () => {
-    const lines: string[] = [];
-    const processor = new PassthroughLogPipe((text) => lines.push(text));
-    processor.push("hello");
-    processor.push("world\nfoo");
-    expect(lines).toEqual(["hello", "world\nfoo"]);
-  });
-
-  it("flush() does nothing", () => {
-    const lines: string[] = [];
-    const processor = new PassthroughLogPipe((text) => lines.push(text));
-    processor.flush();
-    expect(lines).toEqual([]);
-  });
-});
-
-describe("Pymobiledevice3JsonLogPipe", () => {
+describe("Pymobiledevice3JsonLogFilter", () => {
   function jsonLine(overrides: Record<string, unknown> = {}): string {
     return JSON.stringify({
       pid: 67135,
@@ -319,7 +243,7 @@ describe("Pymobiledevice3JsonLogPipe", () => {
     });
   }
 
-  function createPipe(
+  function createFilter(
     options: {
       executableName: string;
       debugDylibOnly?: boolean;
@@ -328,83 +252,60 @@ describe("Pymobiledevice3JsonLogPipe", () => {
       minLevel?: string;
     } = { executableName: "Laboratory" },
   ) {
-    const lines: string[] = [];
-    const pipe = new Pymobiledevice3JsonLogPipe((line) => lines.push(line), options);
-    return { pipe, lines };
+    return new Pymobiledevice3JsonLogFilter(options);
   }
 
   it("keeps app logs and formats them for display", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push(`${jsonLine()}\n`);
-    expect(lines).toEqual([
+    const filter = createFilter();
+    expect(filter.processLine(jsonLine())).toBe(
       "2026-04-16T12:53:59.832625 Laboratory{Laboratory.debug.dylib}[67135] <NOTICE>: ContentView initialized [Laboratory Test][App]",
-    ]);
+    );
   });
 
   it("drops framework noise by image_name basename", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push(`${jsonLine({ image_name: "/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation" })}\n`);
-    expect(lines).toEqual([]);
+    const filter = createFilter();
+    expect(
+      filter.processLine(
+        jsonLine({ image_name: "/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation" }),
+      ),
+    ).toBeNull();
   });
 
   it("passes through non-JSON lines (tunnel notices)", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push("Connected to remote tunnel\n");
-    expect(lines).toEqual(["Connected to remote tunnel"]);
+    const filter = createFilter();
+    expect(filter.processLine("Connected to remote tunnel")).toBe("Connected to remote tunnel");
   });
 
   it("formats entries without a label", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push(`${jsonLine({ label: null })}\n`);
-    expect(lines).toEqual([
+    const filter = createFilter();
+    expect(filter.processLine(jsonLine({ label: null }))).toBe(
       "2026-04-16T12:53:59.832625 Laboratory{Laboratory.debug.dylib}[67135] <NOTICE>: ContentView initialized",
-    ]);
+    );
   });
 
   it("subsystem deny-list drops matching subsystems", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemDenyList: ["com.apple.*"],
     });
-    pipe.push(`${jsonLine({ label: { subsystem: "com.apple.CFBundle", category: "resources" } })}\n`);
-    pipe.push(`${jsonLine()}\n`);
-    expect(lines.length).toBe(1);
-    expect(lines[0]).toContain("ContentView initialized");
+    expect(
+      filter.processLine(jsonLine({ label: { subsystem: "com.apple.CFBundle", category: "resources" } })),
+    ).toBeNull();
+    expect(filter.processLine(jsonLine())).toContain("ContentView initialized");
   });
 
   it("subsystem allow-list keeps only matching subsystems", () => {
-    const { pipe, lines } = createPipe({
+    const filter = createFilter({
       executableName: "Laboratory",
       subsystemAllowList: ["Laboratory Test"],
     });
-    pipe.push(`${jsonLine({ label: { subsystem: "Other", category: "Cat" } })}\n`);
-    pipe.push(`${jsonLine()}\n`);
-    expect(lines.length).toBe(1);
-    expect(lines[0]).toContain("ContentView initialized");
+    expect(filter.processLine(jsonLine({ label: { subsystem: "Other", category: "Cat" } }))).toBeNull();
+    expect(filter.processLine(jsonLine())).toContain("ContentView initialized");
   });
 
   it("honors minimum level", () => {
-    const { pipe, lines } = createPipe({ executableName: "Laboratory", minLevel: "ERROR" });
-    pipe.push(`${jsonLine({ level: "INFO" })}\n`);
-    pipe.push(`${jsonLine({ level: "ERROR" })}\n`);
-    expect(lines.length).toBe(1);
-    expect(lines[0]).toContain("<ERROR>");
-  });
-
-  it("buffers partial chunks", () => {
-    const { pipe, lines } = createPipe();
-    const full = `${jsonLine()}\n`;
-    pipe.push(full.slice(0, 20));
-    expect(lines).toEqual([]);
-    pipe.push(full.slice(20));
-    expect(lines.length).toBe(1);
-  });
-
-  it("flush() emits trailing fragment", () => {
-    const { pipe, lines } = createPipe();
-    pipe.push(jsonLine());
-    expect(lines).toEqual([]);
-    pipe.flush();
-    expect(lines.length).toBe(1);
+    const filter = createFilter({ executableName: "Laboratory", minLevel: "ERROR" });
+    expect(filter.processLine(jsonLine({ level: "INFO" }))).toBeNull();
+    expect(filter.processLine(jsonLine({ level: "ERROR" }))).toContain("<ERROR>");
   });
 });
